server/requirements.txt
django-cors-headers==4.3.1
openai==1.3.6
python-dotenv==1.0.0
